requirements.txt
asgiref==3.2.7
Babel==2.8.0
bandit==1.6.2
coverage==5.1
Django==3.0.6
django-countries==6.1.2
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
drf-extensions==0.6.0
drf-nested-routers==0.91
gitdb==4.0.5
GitPython==3.1.2
pbr==5.4.5
phonenumbers==8.12.4
phonenumberslite==8.12.4
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
smmap==3.0.4
sqlparse==0.3.1
stevedore==1.32.0
